fix(segment): fix type of message ws issue

+++ internal/db/sqlc/models.go
@@ -80,7 +80,7 @@ type InterviewSessions struct {
 type InterviewTurns struct {
 	ID             uuid.UUID      `json:"id"`
 	SessionID      uuid.UUID      `json:"session_id"`
-	TurnNo         int32          `json:"turn_no"`
+	TurnNo         int64          `json:"turn_no"`
 	Actor          string         `json:"actor"`
 	Content        string         `json:"content"`
 	TranscriptText sql.NullString `json:"transcript_text"`
